review - Move ExtensionResolver to core
